네이티브 앱 CLI 빌드 실행 스크립트 추가
