+++ apps/app/android/package.json
@@ -1,0 +1,5 @@
+{
+  "scripts": {
+    "build": "infisical run -- gradle assembleDebug"
+  }
+}
