+++ apps/app/ios/Kosmo/KosmoApp.swift
@@ -2,11 +2,9 @@ import SwiftUI
 
 @main
 struct KosmoApp: App {
-    @State private var model = WebViewModel()
-
     var body: some Scene {
         WindowGroup {
-            WebViewScreen(model: model)
+            WebViewScreen()
                 .ignoresSafeArea()
         }
     }
