+++ apps/app/ios/package.json
@@ -1,0 +1,6 @@
+{
+  "scripts": {
+    "build": "infisical run -- xcodebuild -project Kosmo.xcodeproj -scheme Kosmo -configuration Debug -sdk iphoneos -destination 'generic/platform=iOS' -derivedDataPath build/DerivedData -allowProvisioningUpdates build",
+    "run": "node scripts/run.mjs"
+  }
+}
